Move all pytest fixtures to conftest.py
