+++ tests/runner/conftest.py
@@ -1,8 +1,9 @@
 from random import random
 
+import pandas as pd
 import pytest
 
-from kedro.io import DataCatalog
+from kedro.io import DataCatalog, LambdaDataSet, MemoryDataSet
 from kedro.pipeline import Pipeline, node
 
 
@@ -35,9 +36,52 @@ def return_not_serialisable(arg):  # pylint: disable=unused-argument
     return lambda x: x
 
 
+def multi_input_list_output(arg1, arg2):
+    return [arg1, arg2]
+
+
+@pytest.fixture
+def conflicting_feed_dict(pandas_df_feed_dict):
+    ds1 = MemoryDataSet({"data": 0})
+    ds3 = pandas_df_feed_dict["ds3"]
+    return {"ds1": ds1, "ds3": ds3}
+
+
+@pytest.fixture
+def pandas_df_feed_dict():
+    pandas_df = pd.DataFrame({"Name": ["Alex", "Bob"], "Age": [15, 25]})
+    return {"ds3": pandas_df}
+
+
 @pytest.fixture
 def catalog():
     return DataCatalog()
+
+
+@pytest.fixture
+def memory_catalog():
+    ds1 = MemoryDataSet({"data": 42})
+    ds2 = MemoryDataSet([1, 2, 3, 4, 5])
+    return DataCatalog({"ds1": ds1, "ds2": ds2})
+
+
+@pytest.fixture
+def persistent_dataset_catalog():
+    def _load():
+        return 0
+
+    def _save(arg):
+        assert arg == 0
+
+    persistent_dataset = LambdaDataSet(load=_load, save=_save)
+    return DataCatalog(
+        {
+            "ds0_A": persistent_dataset,
+            "ds0_B": persistent_dataset,
+            "ds2_A": persistent_dataset,
+            "ds2_B": persistent_dataset,
+        }
+    )
 
 
 @pytest.fixture
@@ -86,4 +130,38 @@ def saving_result_pipeline():
 def saving_none_pipeline():
     return Pipeline(
         [node(random, None, "A"), node(return_none, "A", "B"), node(identity, "B", "C")]
+    )
+
+
+@pytest.fixture
+def unfinished_outputs_pipeline():
+    return Pipeline(
+        [
+            node(identity, dict(arg="ds4"), "ds8", name="node1"),
+            node(sink, "ds7", None, name="node2"),
+            node(multi_input_list_output, ["ds3", "ds4"], ["ds6", "ds7"], name="node3"),
+            node(identity, "ds2", "ds5", name="node4"),
+            node(identity, "ds1", "ds4", name="node5"),
+        ]
+    )  # Outputs: ['ds8', 'ds5', 'ds6'] == ['ds1', 'ds2', 'ds3']
+
+
+@pytest.fixture
+def two_branches_crossed_pipeline():
+    """A pipeline with an X-shape (two branches with one common node)"""
+    return Pipeline(
+        [
+            node(identity, "ds0_A", "ds1_A", name="node1_A"),
+            node(identity, "ds0_B", "ds1_B", name="node1_B"),
+            node(
+                multi_input_list_output,
+                ["ds1_A", "ds1_B"],
+                ["ds2_A", "ds2_B"],
+                name="node2",
+            ),
+            node(identity, "ds2_A", "ds3_A", name="node3_A"),
+            node(identity, "ds2_B", "ds3_B", name="node3_B"),
+            node(identity, "ds3_A", "ds4_A", name="node4_A"),
+            node(identity, "ds3_B", "ds4_B", name="node4_B"),
+        ]
     )
